move test_angular_diameter_distance_z1z2
