+++ astropy/cosmology/flrw/tests/test_lambdacdm.py
@@ -23,6 +23,7 @@ from astropy.cosmology.tests.helper import get_redshift_methods
 from astropy.cosmology.tests.test_core import invalid_zs, valid_zs
 from astropy.table import QTable
 from astropy.utils.compat.optional_deps import HAS_SCIPY
+from astropy.utils.exceptions import AstropyUserWarning
 
 from .test_base import FlatFLRWMixinTest, FLRWTest
 
@@ -403,6 +404,55 @@ def test_comoving_transverse_distance_z1z2():
     ) * u.Mpc
 
     assert u.allclose(tcos._comoving_transverse_distance_z1z2(z1, z2), results)
+
+
+@pytest.mark.skipif(not HAS_SCIPY, reason="test requires scipy")
+def test_angular_diameter_distance_z1z2():
+
+    tcos = FlatLambdaCDM(70.4, 0.272, Tcmb0=0.0)
+
+    with pytest.raises(ValueError):  # test diff size z1, z2 fail
+        tcos.angular_diameter_distance_z1z2([1, 2], [3, 4, 5])
+
+    # Tests that should actually work, target values computed with
+    # http://www.astro.multivax.de:8000/phillip/angsiz_prog/README.HTML
+    # Kayser, Helbig, and Schramm (Astron.Astrophys. 318 (1997) 680-686)
+    assert u.allclose(
+        tcos.angular_diameter_distance_z1z2(1, 2), 646.22968662822018 * u.Mpc
+    )
+
+    z1 = 2  # Separate test for z2<z1, returns negative value with warning
+    z2 = 1
+    results = -969.34452994 * u.Mpc
+    with pytest.warns(AstropyUserWarning, match="less than first redshift"):
+        assert u.allclose(tcos.angular_diameter_distance_z1z2(z1, z2), results)
+
+    z1 = 0, 0, 0.5, 1
+    z2 = 2, 1, 2.5, 1.1
+    results = (
+        1760.0628637762106,
+        1670.7497657219858,
+        1159.0970895962193,
+        115.72768186186921,
+    ) * u.Mpc
+
+    assert u.allclose(tcos.angular_diameter_distance_z1z2(z1, z2), results)
+
+    z1 = 0.1
+    z2 = 0.1, 0.2, 0.5, 1.1, 2
+    results = (0.0, 332.09893173, 986.35635069, 1508.37010062, 1621.07937976) * u.Mpc
+    assert u.allclose(tcos.angular_diameter_distance_z1z2(0.1, z2), results)
+
+    # Non-flat (positive Ok0) test
+    tcos = LambdaCDM(H0=70.4, Om0=0.2, Ode0=0.5, Tcmb0=0.0)
+    assert u.allclose(
+        tcos.angular_diameter_distance_z1z2(1, 2), 620.1175337852428 * u.Mpc
+    )
+    # Non-flat (negative Ok0) test
+    tcos = LambdaCDM(H0=100, Om0=2, Ode0=1, Tcmb0=0.0)
+    assert u.allclose(
+        tcos.angular_diameter_distance_z1z2(1, 2), 228.42914659246014 * u.Mpc
+    )
 
 
 ##############################################################################
